extract CLI configuration from main.go
